libretro: Update core options API
